+++ codec/hessian/decode.go
@@ -16,12 +16,13 @@ import (
 	"io"
 	"math"
 	"reflect"
+	"strconv"
 	"strings"
 	"time"
 )
 
 import (
-	"github.com/pkg/errors"
+	jerrors "github.com/juju/errors"
 )
 
 type Decoder struct {
@@ -115,7 +116,7 @@ func (d *Decoder) decType() (string, error) {
 
 	buf = arr[:1]
 	if _, err = io.ReadFull(d.reader, buf); err != nil {
-		return "", newCodecError("decType reading tag", err)
+		return "", jerrors.Annotate(err, "decType reading tag")
 	}
 	tag = buf[0]
 	if (tag >= BC_STRING_DIRECT && tag <= STRING_DIRECT_MAX) ||
@@ -124,7 +125,7 @@ func (d *Decoder) decType() (string, error) {
 	}
 
 	if idx, err = d.decInt32(TAG_READ); err != nil {
-		return "", newCodecError("decType reading tag", err)
+		return "", jerrors.Annotate(err, "decType reading tag")
 	}
 
 	typ, _, err = d.getStructDefByIndex(int(idx))
@@ -139,69 +140,69 @@ func (d *Decoder) decType() (string, error) {
 func (d *Decoder) Decode() (interface{}, error) {
 	var (
 		err error
-		t   byte
+		tag byte
 	)
 
-	t, err = d.readByte()
+	tag, err = d.readByte()
 	if err == io.EOF {
 		return nil, err
 	}
 	switch {
-	case t == BC_END:
+	case tag == BC_END:
 		return nil, nil
 
-	case t == BC_NULL: // 'N': //null
+	case tag == BC_NULL: // 'N': //null
 		return nil, nil
 
-	case t == BC_TRUE: // 'T': //true
+	case tag == BC_TRUE: // 'T': //true
 		return true, nil
 
-	case t == BC_FALSE: //'F': //false
+	case tag == BC_FALSE: //'F': //false
 		return false, nil
 
-	case (0x80 <= t && t <= 0xbf) || (0xc0 <= t && t <= 0xcf) ||
-		(0xd0 <= t && t <= 0xd7) || t == BC_INT: //'I': //int
-		return d.decInt32(int32(t))
+	case (0x80 <= tag && tag <= 0xbf) || (0xc0 <= tag && tag <= 0xcf) ||
+		(0xd0 <= tag && tag <= 0xd7) || tag == BC_INT: //'I': //int
+		return d.decInt32(int32(tag))
 
-	case (t >= 0xd8 && t <= 0xef) || (t >= 0xf4 && t <= 0xff) ||
-		(t >= 0x38 && t <= 0x3f) || (t == BC_LONG_INT) || (t == BC_LONG): //'L': //long
-		return d.decInt64(int32(t))
+	case (tag >= 0xd8 && tag <= 0xef) || (tag >= 0xf0 && tag <= 0xff) ||
+		(tag >= 0x38 && tag <= 0x3f) || (tag == BC_LONG_INT) || (tag == BC_LONG): //'L': //long
+		return d.decInt64(int32(tag))
 
-	case (t == BC_DATE_MINUTE) || (t == BC_DATE): //'d': //date
-		return d.decDate(int32(t))
+	case (tag == BC_DATE_MINUTE) || (tag == BC_DATE): //'d': //date
+		return d.decDate(int32(tag))
 
-	case (t == BC_DOUBLE_ZERO) || (t == BC_DOUBLE_ONE) || (t == BC_DOUBLE_BYTE) ||
-		(t == BC_DOUBLE_SHORT) || (t == BC_DOUBLE_MILL) || (t == BC_DOUBLE): //'D': //double
-		return d.decDouble(int32(t))
+	case (tag == BC_DOUBLE_ZERO) || (tag == BC_DOUBLE_ONE) || (tag == BC_DOUBLE_BYTE) ||
+		(tag == BC_DOUBLE_SHORT) || (tag == BC_DOUBLE_MILL) || (tag == BC_DOUBLE): //'D': //double
+		return d.decDouble(int32(tag))
 
 	// case 'S', 's', 'X', 'x': //string,xml
-	case (t == BC_STRING_CHUNK || t == BC_STRING) ||
-		(t >= BC_STRING_DIRECT && t <= STRING_DIRECT_MAX) ||
-		(t >= 0x30 && t <= 0x33):
-		return d.decString(int32(t))
+	case (tag == BC_STRING_CHUNK || tag == BC_STRING) ||
+		(tag >= BC_STRING_DIRECT && tag <= STRING_DIRECT_MAX) ||
+		(tag >= 0x30 && tag <= 0x33):
+		return d.decString(int32(tag))
 
 		// case 'B', 'b': //binary
-	case (t == BC_BINARY) || (t == BC_BINARY_CHUNK) || (t >= 0x20 && t <= 0x2f):
-		return d.decBinary(int32(t))
+	case (tag == BC_BINARY) || (tag == BC_BINARY_CHUNK) || (tag >= 0x20 && tag <= 0x2f):
+		return d.decBinary(int32(tag))
 
 	// case 'V': //list
-	case (t >= BC_LIST_DIRECT && t <= 0x77) || (t == BC_LIST_FIXED || t == BC_LIST_VARIABLE) ||
-		(t >= BC_LIST_DIRECT_UNTYPED && t <= 0x7f) ||
-		(t == BC_LIST_FIXED_UNTYPED || t == BC_LIST_VARIABLE_UNTYPED):
-		return d.decList(int32(t))
+	case (tag >= BC_LIST_DIRECT && tag <= 0x77) || (tag == BC_LIST_FIXED || tag == BC_LIST_VARIABLE) ||
+		(tag >= BC_LIST_DIRECT_UNTYPED && tag <= 0x7f) ||
+		(tag == BC_LIST_FIXED_UNTYPED || tag == BC_LIST_VARIABLE_UNTYPED):
+		return d.decList(int32(tag))
 
-	case (t == BC_MAP) || (t == BC_MAP_UNTYPED):
-		return d.decMap(int32(t))
+	case (tag == BC_MAP) || (tag == BC_MAP_UNTYPED):
+		return d.decMap(int32(tag))
 
-	case (t == BC_OBJECT_DEF) || (t == BC_OBJECT) ||
-		(BC_OBJECT_DIRECT <= t && t <= (BC_OBJECT_DIRECT+OBJECT_DIRECT_MAX)):
-		return d.decObject(int32(t))
+	case (tag == BC_OBJECT_DEF) || (tag == BC_OBJECT) ||
+		(BC_OBJECT_DIRECT <= tag && tag <= (BC_OBJECT_DIRECT+OBJECT_DIRECT_MAX)):
+		return d.decObject(int32(tag))
 
-	case (t == BC_REF): // 'R': //ref, 一个整数，用以指代前面的list 或者 map
-		return d.decRef(int32(t))
+	case (tag == BC_REF): // 'R': //ref, 一个整数，用以指代前面的list 或者 map
+		return d.decRef(int32(tag))
 
 	default:
-		return nil, fmt.Errorf("Invalid type: %v,>>%v<<<", string(t), d.peek(d.len()))
+		return nil, fmt.Errorf("Invalid type: %v,>>%v<<<", string(tag), d.peek(d.len()))
 	}
 }
 
@@ -234,24 +235,24 @@ func (d *Decoder) decInt32(flag int32) (int32, error) {
 
 	case tag >= 0xc0 && tag <= 0xcf:
 		if _, err = io.ReadFull(d.reader, buf[:1]); err != nil {
-			return 0, newCodecError("decInt32 short integer", err)
+			return 0, jerrors.Annotate(err, "decInt32 short integer")
 		}
 		return int32(tag-BC_INT_BYTE_ZERO)<<8 + int32(buf[0]), nil
 
 	case tag >= 0xd0 && tag <= 0xd7:
 		if _, err = io.ReadFull(d.reader, buf[:2]); err != nil {
-			return 0, newCodecError("decInt32 short integer", err)
+			return 0, jerrors.Annotate(err, "decInt32 short integer")
 		}
 		return int32(tag-BC_INT_SHORT_ZERO)<<16 + int32(buf[0])<<8 + int32(buf[1]), nil
 
 	case tag == BC_INT:
 		if _, err := io.ReadFull(d.reader, buf[:4]); err != nil {
-			return 0, newCodecError("decInt32 parse int", err)
+			return 0, jerrors.Annotate(err, "decInt32 parse int")
 		}
 		return int32(buf[0])<<24 + int32(buf[1])<<16 + int32(buf[2])<<8 + int32(buf[3]), nil
 
 	default:
-		return 0, newCodecError("decInt32 integer wrong tag:" + fmt.Sprintf("%d-%#x", int(tag), tag))
+		return 0, jerrors.Errorf("decInt32 integer wrong tag:%d-%#x", int(tag), tag)
 	}
 }
 
@@ -265,6 +266,7 @@ func (d *Decoder) decInt32(flag int32) (int32, error) {
 // ::= [xf0-xff] b0          # -x800 to x7ff
 // ::= [x38-x3f] b1 b0       # -x40000 to x3ffff
 // ::= x59 b3 b2 b1 b0       # 32-bit integer cast to long
+// ref: hessian-lite/src/main/java/com/alibaba/com/caucho/hessian/io/Hessian2Input.java:readLong
 func (d *Decoder) decInt64(flag int32) (int64, error) {
 	var (
 		err error
@@ -279,31 +281,90 @@ func (d *Decoder) decInt64(flag int32) (int64, error) {
 	}
 
 	switch {
+	case tag == byte('N'):
+		return int64(0), nil
+
+	case tag == byte('F'):
+		return int64(0), nil
+
+	case tag == byte('T'):
+		return int64(1), nil
+
+		// direct integer
+	case tag >= 0x80 && tag <= 0xbf:
+		return int64(tag - BC_INT_ZERO), nil
+
+		// byte int
+	case tag >= 0xc0 && tag <= 0xcf:
+		if _, err = io.ReadFull(d.reader, buf[:1]); err != nil {
+			return 0, jerrors.Annotate(err, "decInt64 short integer")
+		}
+		return int64(tag-BC_INT_BYTE_ZERO)<<8 + int64(buf[0]), nil
+
+		// short int
+	case tag >= 0xd0 && tag <= 0xd7:
+		if _, err = io.ReadFull(d.reader, buf[:2]); err != nil {
+			return 0, jerrors.Annotate(err, "decInt64 short integer")
+		}
+		return int64(tag-BC_INT_SHORT_ZERO)<<16 + int64(buf[0])<<8 + int64(buf[1]), nil
+
+	case tag == BC_DOUBLE_BYTE:
+		tag, _ = d.readByte()
+		return int64(tag), nil
+
+	case tag == BC_DOUBLE_SHORT:
+		if _, err = io.ReadFull(d.reader, buf[:2]); err != nil {
+			return 0, jerrors.Annotate(err, "decInt64 short integer")
+		}
+
+		return int64(int(buf[0])<<8 + int(buf[1])), nil
+
+	case tag == BC_INT: // 'I'
+		i32, err := d.decInt32(TAG_READ)
+		return int64(i32), err
+
+	case tag == BC_LONG_INT: // x59
+		i32, err := d.decInt32(TAG_READ)
+		return int64(i32), err
+
+		// direct long
 	case tag >= 0xd8 && tag <= 0xef:
 		return int64(tag - BC_LONG_ZERO), nil
 
-	case tag >= 0xf4 && tag <= 0xff:
+		// byte long
+	case tag >= 0xf0 && tag <= 0xff:
 		if _, err = io.ReadFull(d.reader, buf[:1]); err != nil {
-			return 0, newCodecError("decInt64 short integer", err)
+			return 0, jerrors.Annotate(err, "decInt64 short integer")
 		}
 		return int64(tag-BC_LONG_BYTE_ZERO)<<8 + int64(buf[0]), nil
 
-	case tag >= 0x38 && tag <= 0x3f:
+		// short long
+	case tag >= 0x38 && tag <= 0x3f: // ['8',  '?']
 		if _, err := io.ReadFull(d.reader, buf[:2]); err != nil {
-			return 0, newCodecError("decInt64 short integer", err)
+			return 0, jerrors.Annotate(err, "decInt64 short integer")
 		}
 		return int64(tag-BC_LONG_SHORT_ZERO)<<16 + int64(buf[0])<<8 + int64(buf[1]), nil
 		// return int64(tag-BC_LONG_SHORT_ZERO)<<16 + int64(buf[0])*256 + int64(buf[1]), nil
 
-	case tag == BC_LONG:
+	case tag == BC_LONG: // 'L'
 		if _, err := io.ReadFull(d.reader, buf[:8]); err != nil {
-			return 0, newCodecError("decInt64 parse long", err)
+			return 0, jerrors.Annotate(err, "decInt64 parse long")
 		}
 		return int64(buf[0])<<56 + int64(buf[1])<<48 + int64(buf[2])<<40 + int64(buf[3])<<32 +
 			int64(buf[4])<<24 + int64(buf[5])<<16 + int64(buf[6])<<8 + int64(buf[7]), nil
 
+	case tag == BC_DOUBLE_ZERO:
+		return int64(0), nil
+
+	case tag == BC_DOUBLE_ONE:
+		return int64(1), nil
+
+	case tag == BC_DOUBLE_MILL:
+		i64, err := d.decInt32(TAG_READ)
+		return int64(i64), jerrors.Annotate(err, "decInt32")
+
 	default:
-		return 0, newCodecError("decInt64 long wrong tag " + fmt.Sprintf("%d-%#x", int(tag), tag))
+		return 0, jerrors.Errorf("decInt64 long wrong tag:%d-%#x", int(tag), tag)
 	}
 }
 
@@ -401,7 +462,7 @@ func (d *Decoder) decDouble(flag int32) (interface{}, error) {
 
 	case BC_DOUBLE_SHORT:
 		if _, err = io.ReadFull(d.reader, buf[:2]); err != nil {
-			return nil, newCodecError("decDouble short integer", err)
+			return nil, jerrors.Annotate(err, "decDouble short integer")
 		}
 
 		return float64(int(buf[0])<<8 + int(buf[1])), nil
@@ -412,7 +473,7 @@ func (d *Decoder) decDouble(flag int32) (interface{}, error) {
 
 	case BC_DOUBLE:
 		if _, err = io.ReadFull(d.reader, buf[:8]); err != nil {
-			return nil, newCodecError("decDouble short integer", err)
+			return nil, jerrors.Annotate(err, "decDouble short integer")
 		}
 
 		bits := binary.BigEndian.Uint64(buf[:8])
@@ -420,7 +481,7 @@ func (d *Decoder) decDouble(flag int32) (interface{}, error) {
 		return datum, nil
 	}
 
-	return nil, newCodecError("decDouble parse double wrong tag:" + fmt.Sprintf("%d-%#x", int(tag), tag))
+	return nil, jerrors.Errorf("decDouble parse double wrong tag:%d-%#x", int(tag), tag)
 }
 
 /////////////////////////////////////////
@@ -446,7 +507,7 @@ func (d *Decoder) getStrLen(tag byte) (int32, error) {
 	case tag >= 0x30 && tag <= 0x33:
 		_, err = io.ReadFull(d.reader, buf[:1])
 		if err != nil {
-			return -1, newCodecError("getStrLen byte4 integer", err)
+			return -1, jerrors.Annotate(err, "getStrLen byte4 integer")
 		}
 
 		length = int32(tag-0x30)<<8 + int32(buf[0])
@@ -455,13 +516,13 @@ func (d *Decoder) getStrLen(tag byte) (int32, error) {
 	case tag == BC_STRING_CHUNK || tag == BC_STRING:
 		_, err = io.ReadFull(d.reader, buf[:2])
 		if err != nil {
-			return -1, newCodecError("getStrLen byte5 integer", err)
+			return -1, jerrors.Annotate(err, "getStrLen byte5 integer")
 		}
 		length = int32(buf[0])<<8 + int32(buf[1])
 		return length, nil
 
 	default:
-		return -1, newCodecError("getStrLen getStrLen")
+		return -1, jerrors.Annotate(err, "getStrLen getStrLen")
 	}
 }
 
@@ -473,24 +534,25 @@ func getRune(reader io.Reader) (rune, int, error) {
 
 	typ = reflect.TypeOf(reader.(interface{}))
 
-	if (typ == reflect.TypeOf(&bufio.Reader{})) {
+	switch {
+	case typ == reflect.TypeOf(&bufio.Reader{}):
 		byteReader := reader.(interface{}).(*bufio.Reader)
 		return byteReader.ReadRune()
-	}
 
-	if (typ == reflect.TypeOf(&bytes.Buffer{})) {
+	case typ == reflect.TypeOf(&bytes.Buffer{}):
 		byteReader := reader.(interface{}).(*bytes.Buffer)
 		return byteReader.ReadRune()
-	}
 
-	if (typ == reflect.TypeOf(&bytes.Reader{})) {
+	case typ == reflect.TypeOf(&bytes.Reader{}):
 		byteReader := reader.(interface{}).(*bytes.Reader)
 		return byteReader.ReadRune()
-	}
 
-	return runeNil, 0, nil
+	default:
+		return runeNil, 0, nil
+	}
 }
 
+// hessian-lite/src/main/java/com/alibaba/com/caucho/hessian/io/Hessian2Input.java : readString
 func (d *Decoder) decString(flag int32) (string, error) {
 	var (
 		tag    byte
@@ -506,6 +568,45 @@ func (d *Decoder) decString(flag int32) (string, error) {
 		tag, _ = d.readByte()
 	}
 
+	switch {
+	case tag == byte(BC_NULL):
+		return "null", nil
+
+	case tag == byte(BC_TRUE):
+		return "true", nil
+
+	case tag == byte(BC_FALSE):
+		return "false", nil
+
+	case (0x80 <= tag && tag <= 0xbf) || (0xc0 <= tag && tag <= 0xcf) ||
+		(0xd0 <= tag && tag <= 0xd7) || tag == BC_INT ||
+		(tag >= 0xd8 && tag <= 0xef) || (tag >= 0xf0 && tag <= 0xff) ||
+		(tag >= 0x38 && tag <= 0x3f) || (tag == BC_LONG_INT) || (tag == BC_LONG):
+		i64, err := d.decInt64(int32(tag))
+		if err != nil {
+			return "", jerrors.Annotatef(err, "tag:%+v", tag)
+		}
+
+		return strconv.Itoa(int(i64)), nil
+
+	case tag == byte(BC_DOUBLE_ZERO):
+		return "0.0", nil
+
+	case tag == byte(BC_DOUBLE_ONE):
+		return "1.0", nil
+
+	case tag == byte(BC_DOUBLE_BYTE) || tag == byte(BC_DOUBLE_SHORT):
+		f, err := d.decDouble(int32(tag))
+		if err != nil {
+			return "", jerrors.Annotatef(err, "tag:%+v", tag)
+		}
+
+		return strconv.FormatFloat(f.(float64), 'E', -1, 64), nil
+		//var fs string
+		//fmt.Sprintf(fs, "%f", f.(float64))
+		//return fs, nil
+	}
+
 	last = true
 	if (tag >= BC_STRING_DIRECT && tag <= STRING_DIRECT_MAX) ||
 		(tag >= 0x30 && tag <= 0x33) ||
@@ -519,7 +620,7 @@ func (d *Decoder) decString(flag int32) (string, error) {
 
 		l, err := d.getStrLen(tag)
 		if err != nil {
-			return s, newCodecError("decString->getStrLen", err)
+			return s, jerrors.Annotate(err, "decString->getStrLen")
 		}
 		length = l
 		runeDate := make([]rune, length)
@@ -543,7 +644,7 @@ func (d *Decoder) decString(flag int32) (string, error) {
 
 					l, err := d.getStrLen(b)
 					if err != nil {
-						return s, newCodecError("decString->getStrLen", err)
+						return s, jerrors.Annotate(err, "decString->getStrLen")
 					}
 					length += l
 					bs := make([]rune, length)
@@ -551,13 +652,13 @@ func (d *Decoder) decString(flag int32) (string, error) {
 					runeDate = bs
 
 				default:
-					return s, newCodecError("decString tag error ", err)
+					return s, jerrors.Annotate(err, "decString tag")
 				}
 
 			} else {
 				r, _, err = d.reader.ReadRune()
 				if err != nil {
-					return s, newCodecError("decString->ReadRune", err)
+					return s, jerrors.Annotate(err, "decString->ReadRune")
 				}
 				runeDate[i] = r
 				i++
@@ -567,7 +668,7 @@ func (d *Decoder) decString(flag int32) (string, error) {
 		return string(runeDate), nil
 	}
 
-	return s, newCodecError(fmt.Sprintf("unknown string tag %#x\n", tag))
+	return s, jerrors.Errorf("unknown string tag %#x\n", tag)
 }
 
 /////////////////////////////////////////
@@ -590,7 +691,7 @@ func (d *Decoder) getBinaryLength(tag byte) (int, error) {
 	}
 
 	if _, err = io.ReadFull(d.reader, buf[:2]); err != nil {
-		return 0, newCodecError("getBinaryLength parse binary", err)
+		return 0, jerrors.Annotate(err, "getBinaryLength parse binary")
 	}
 
 	return int(buf[0]<<8 + buf[1]), nil
@@ -619,7 +720,7 @@ func (d *Decoder) decBinary(flag int32) ([]byte, error) {
 		}
 		l, err := d.getBinaryLength(tag)
 		if err != nil {
-			return nil, newCodecError("decBinary->getBinaryLength", err)
+			return nil, jerrors.Annotate(err, "decBinary->getBinaryLength")
 		}
 		length = int32(l)
 		data := make([]byte, length)
@@ -632,7 +733,7 @@ func (d *Decoder) decBinary(flag int32) ([]byte, error) {
 				var buf [1]byte
 				_, err := io.ReadFull(d.reader, buf[:1])
 				if err != nil {
-					return nil, newCodecError("decBinary byte1 integer", err)
+					return nil, jerrors.Annotate(err, "decBinary byte1 integer")
 				}
 				b := buf[0]
 				switch {
@@ -644,20 +745,20 @@ func (d *Decoder) decBinary(flag int32) ([]byte, error) {
 					}
 					l, err := d.getStrLen(b)
 					if err != nil {
-						return nil, newCodecError("decBinary getStrLen", err)
+						return nil, jerrors.Annotate(err, "decBinary getStrLen")
 					}
 					length += l
 					bs := make([]byte, 0, length)
 					copy(bs, data)
 					data = bs
 				default:
-					return nil, newCodecError("decBinary tag error ", err)
+					return nil, jerrors.Annotate(err, "decBinary tag error")
 				}
 			} else {
 				var buf [1]byte
 				_, err := io.ReadFull(d.reader, buf[:1])
 				if err != nil {
-					return nil, newCodecError("decBinary byte2 integer", err)
+					return nil, jerrors.Annotate(err, "decBinary byte2 integer")
 				}
 				data[i] = buf[0]
 				i++
@@ -665,9 +766,9 @@ func (d *Decoder) decBinary(flag int32) ([]byte, error) {
 		}
 
 		return data, nil
-	} else {
-		return nil, newCodecError("decBinary byte3 integer")
 	}
+
+	return nil, jerrors.Errorf("decBinary byte3 integer")
 }
 
 /////////////////////////////////////////
@@ -690,7 +791,7 @@ func (d *Decoder) readBufByte() (byte, error) {
 
 	_, err = io.ReadFull(d.reader, buf[:1])
 	if err != nil {
-		return 0, newCodecError("readBufByte", err)
+		return 0, jerrors.Annotate(err, "readBufByte")
 	}
 
 	return buf[0], nil
@@ -708,7 +809,7 @@ func (d *Decoder) decSlice(value reflect.Value) (interface{}, error) {
 	} else {
 		ii, err := d.decInt32(TAG_READ)
 		if err != nil {
-			return nil, newCodecError("decSlice->decInt32", err)
+			return nil, jerrors.Annotate(err, "decSlice->decInt32")
 		}
 		i = int(ii)
 	}
@@ -717,7 +818,7 @@ func (d *Decoder) decSlice(value reflect.Value) (interface{}, error) {
 	for j := 0; j < i; j++ {
 		it, err := d.Decode()
 		if err != nil {
-			return nil, newCodecError("decSlice->ReadList", err)
+			return nil, jerrors.Annotate(err, "decSlice->ReadList")
 		}
 		ary.Index(j).Set(reflect.ValueOf(it))
 	}
@@ -762,7 +863,7 @@ func (d *Decoder) decList(flag int32) (interface{}, error) {
 	case (tag >= BC_LIST_DIRECT && tag <= 0x77) || (tag == BC_LIST_FIXED || tag == BC_LIST_VARIABLE):
 		// str, err := d.decType()
 		// if err != nil {
-		// 	return nil, newCodecError("ReadType", err)
+		// 	return nil, jerrors.Annotate(err,"ReadType", err)
 		// }
 		d.decType() // 忽略
 		if tag >= BC_LIST_DIRECT && tag <= 0x77 {
@@ -770,7 +871,7 @@ func (d *Decoder) decList(flag int32) (interface{}, error) {
 		} else {
 			i32, err := d.decInt32(TAG_READ)
 			if err != nil {
-				return nil, newCodecError("decList->decInt32", err)
+				return nil, jerrors.Annotate(err, "decList->decInt32")
 			}
 			size = int(i32)
 		}
@@ -780,7 +881,7 @@ func (d *Decoder) decList(flag int32) (interface{}, error) {
 		for j := 0; j < size; j++ {
 			it, err := d.Decode()
 			if err != nil {
-				return nil, newCodecError("decList->Decode", err)
+				return nil, jerrors.Annotate(err, "decList->Decode")
 			}
 			ary[j] = it
 		}
@@ -797,7 +898,7 @@ func (d *Decoder) decList(flag int32) (interface{}, error) {
 		} else {
 			i32, err := d.decInt32(TAG_READ)
 			if err != nil {
-				return nil, newCodecError("decList->ReadType", err)
+				return nil, jerrors.Annotate(err, "decList->ReadType")
 			}
 			size = int(i32)
 		}
@@ -806,7 +907,7 @@ func (d *Decoder) decList(flag int32) (interface{}, error) {
 		for j := 0; j < size; j++ {
 			it, err := d.Decode()
 			if err != nil {
-				return nil, newCodecError("decList->Decode", err)
+				return nil, jerrors.Annotate(err, "decList->Decode")
 			}
 			ary[j] = it
 		}
@@ -818,7 +919,7 @@ func (d *Decoder) decList(flag int32) (interface{}, error) {
 		return ary, nil
 
 	default:
-		return nil, newCodecError("illegal list type tag:", tag)
+		return nil, jerrors.Errorf("illegal list type tag:%+v", tag)
 	}
 }
 
@@ -839,7 +940,7 @@ func (d *Decoder) decMapByValue(value reflect.Value) (interface{}, error) {
 	} else if tag == BC_MAP_UNTYPED {
 		//do nothing
 	} else {
-		return nil, newCodecError("wrong header BC_MAP_UNTYPED")
+		return nil, jerrors.Errorf("wrong header BC_MAP_UNTYPED")
 	}
 
 	m := reflect.MakeMap(value.Type())
@@ -850,7 +951,7 @@ func (d *Decoder) decMapByValue(value reflect.Value) (interface{}, error) {
 			if err == io.EOF {
 				break
 			} else {
-				return nil, newCodecError("decMapByValue->ReadType", err)
+				return nil, jerrors.Annotate(err, "decMapByValue->ReadType")
 			}
 		}
 		if key == nil {
@@ -967,7 +1068,7 @@ func (d *Decoder) decMap(flag int32) (interface{}, error) {
 		return m, nil
 
 	default:
-		return nil, newCodecError("illegal map type tag:", tag)
+		return nil, jerrors.Errorf("illegal map type tag:%+v", tag)
 	}
 }
 
@@ -1051,17 +1152,17 @@ func (d *Decoder) decClassDef() (interface{}, error) {
 
 	clsName, err = d.decString(TAG_READ)
 	if err != nil {
-		return nil, newCodecError("decClassDef->decString", err)
+		return nil, jerrors.Annotate(err, "decClassDef->decString, to get class name")
 	}
 	fieldNum, err = d.decInt32(TAG_READ)
 	if err != nil {
-		return nil, newCodecError("decClassDef->decInt32", err)
+		return nil, jerrors.Annotate(err, "decClassDef->decInt32, to get field num")
 	}
 	fieldList = make([]string, fieldNum)
 	for i := 0; i < int(fieldNum); i++ {
 		fieldName, err = d.decString(TAG_READ)
 		if err != nil {
-			return nil, newCodecError("decClassDef->decString", err)
+			return nil, jerrors.Annotatef(err, "decClassDef->decString, filed num:%d, index:%d", fieldNum, i)
 		}
 		fieldList[i] = fieldName
 	}
@@ -1081,7 +1182,8 @@ func findField(name string, typ reflect.Type) (int, error) {
 			return i, nil
 		}
 	}
-	return 0, newCodecError("findField")
+
+	return 0, jerrors.Errorf("failed to find field %s", name)
 }
 
 func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, error) {
@@ -1091,7 +1193,7 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 	)
 
 	if typ.Kind() != reflect.Struct {
-		return nil, newCodecError("wrong type expect Struct but get " + typ.String())
+		return nil, jerrors.Errorf("wrong type expect Struct but get:%s", typ.String())
 	}
 	vv := reflect.New(typ)
 	st := reflect.ValueOf(vv.Interface()).Elem()
@@ -1106,7 +1208,7 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 		fldValue := st.Field(index)
 		// fmt.Println("fld", fldName, fldValue, fldValue.Kind())
 		if !fldValue.CanSet() {
-			return nil, newCodecError("decInstance CanSet false for " + fldName)
+			return nil, jerrors.Errorf("decInstance CanSet false for field %s", fldName)
 		}
 
 		kind := fldValue.Kind()
@@ -1116,7 +1218,7 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 		case kind == reflect.String:
 			str, err := d.decString(TAG_READ)
 			if err != nil {
-				return nil, newCodecError("decInstance->ReadString:"+fldName, err)
+				return nil, jerrors.Annotatef(err, "decInstance->ReadString: %s", fldName)
 			}
 			fldValue.SetString(str)
 
@@ -1128,12 +1230,12 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 					d.unreadByte() // enum解析，上面decInt64已经读取一个字节，所以这里需要回退一个字节
 					s, err := d.Decode()
 					if err != nil {
-						return nil, newCodecError("decInstance->decObject:"+fldName, err)
+						return nil, jerrors.Annotatef(err, "decInstance->decObject field name:%s", fldName)
 					}
 					enumValue, _ := s.(JavaEnum)
 					num = int32(enumValue)
 				} else {
-					return nil, newCodecError("decInstance->ParseInt:"+fldName, err)
+					return nil, jerrors.Annotatef(err, "decInstance->ParseInt, field name:%s", fldName)
 				}
 			}
 
@@ -1146,12 +1248,12 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 					d.unreadByte() // enum解析，上面decInt64已经读取一个字节，所以这里需要回退一个字节
 					s, err := d.Decode()
 					if err != nil {
-						return nil, newCodecError("decInstance->decObject:"+fldName, err)
+						return nil, jerrors.Annotatef(err, "decInstance->decObject field name:%s", fldName)
 					}
 					enumValue, _ := s.(JavaEnum)
 					num = int64(enumValue)
 				} else {
-					return nil, newCodecError("decInstance->decInt64:"+fldName, err)
+					return nil, jerrors.Annotatef(err, "decInstance->decInt64 field name:%s", fldName)
 				}
 			}
 
@@ -1160,14 +1262,14 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 		case kind == reflect.Bool:
 			b, err := d.Decode()
 			if err != nil {
-				return nil, newCodecError("decInstance->Decode:"+fldName, err)
+				return nil, jerrors.Annotatef(err, "decInstance->Decode field name:%s", fldName)
 			}
 			fldValue.SetBool(b.(bool))
 
 		case kind == reflect.Float32 || kind == reflect.Float64:
 			num, err := d.decDouble(TAG_READ)
 			if err != nil {
-				return nil, newCodecError("decInstance->decDouble"+fldName, err)
+				return nil, jerrors.Annotatef(err, "decInstance->decDouble field name:%s", fldName)
 			}
 			fldValue.SetFloat(num.(float64))
 
@@ -1194,13 +1296,13 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 			if fldValue.Type().String() == "time.Time" {
 				s, err = d.decDate(TAG_READ)
 				if err != nil {
-					return nil, newCodecError("decInstance->decDate", err)
+					return nil, jerrors.Annotate(err, "decInstance->decDate")
 				}
 				fldValue.Set(reflect.ValueOf(s))
 			} else {
 				s, err = d.decObject(TAG_READ)
 				if err != nil {
-					return nil, newCodecError("decInstance->decObject", err)
+					return nil, jerrors.Annotate(err, "decInstance->decObject")
 				}
 				if s != nil {
 					fldValue.Set(reflect.Indirect(s.(reflect.Value)))
@@ -1208,7 +1310,7 @@ func (d *Decoder) decInstance(typ reflect.Type, cls classDef) (interface{}, erro
 			}
 
 		default:
-			return nil, fmt.Errorf("unknown struct member type:%d", kind)
+			return nil, jerrors.Errorf("unknown struct member type:%d", kind)
 		}
 	}
 
@@ -1228,12 +1330,12 @@ func (d *Decoder) getStructDefByIndex(idx int) (reflect.Type, classDef, error) {
 	)
 
 	if len(d.clsDefList) <= idx || idx < 0 {
-		return nil, cls, fmt.Errorf("illegal class index @idx %d", idx)
+		return nil, cls, jerrors.Errorf("illegal class index @idx %d", idx)
 	}
 	cls = d.clsDefList[idx]
 	s, ok = getStructInfo(cls.javaName)
 	if !ok {
-		return nil, cls, fmt.Errorf("can not find go type name %s in registry", clsName)
+		return nil, cls, jerrors.Errorf("can not find go type name %s in registry", clsName)
 	}
 
 	return s.typ, cls, nil
@@ -1249,11 +1351,11 @@ func (d *Decoder) decEnum(javaName string, flag int32) (JavaEnum, error) {
 	)
 	enumName, err = d.decString(TAG_READ) // java enum class member is "name"
 	if err != nil {
-		return InvalidJavaEnum, errors.Wrapf(err, "decString for decJavaEnum")
+		return InvalidJavaEnum, jerrors.Annotate(err, "decString for decJavaEnum")
 	}
 	info, ok = getStructInfo(javaName)
 	if !ok {
-		return InvalidJavaEnum, fmt.Errorf("getStructInfo(javaName:%s) = false", javaName)
+		return InvalidJavaEnum, jerrors.Errorf("getStructInfo(javaName:%s) = false", javaName)
 	}
 
 	enumValue = info.inst.(POJOEnum).EnumValue(enumName)
@@ -1280,7 +1382,7 @@ func (d *Decoder) decObject(flag int32) (interface{}, error) {
 	case tag == BC_OBJECT_DEF:
 		clsDef, err := d.decClassDef()
 		if err != nil {
-			return nil, newCodecError("decObject->decClassDef byte double", err)
+			return nil, jerrors.Annotate(err, "decObject->decClassDef byte double")
 		}
 		cls, _ = clsDef.(classDef)
 		//add to slice
@@ -1316,7 +1418,7 @@ func (d *Decoder) decObject(flag int32) (interface{}, error) {
 		return d.decInstance(typ, cls)
 
 	default:
-		return nil, newCodecError("decObject illegal object type tag:", tag)
+		return nil, jerrors.Errorf("decObject illegal object type tag:%+v", tag)
 	}
 }
 
@@ -1352,6 +1454,6 @@ func (d *Decoder) decRef(flag int32) (interface{}, error) {
 		return &d.refs[i], nil
 
 	default:
-		return nil, newCodecError("decRef illegal ref type tag:", tag)
+		return nil, jerrors.Errorf("decRef illegal ref type tag:%+v", tag)
 	}
 }