+++ codec/hessian/encode.go
@@ -19,6 +19,7 @@ import (
 
 import (
 	"github.com/AlexStocks/goext/strings"
+	jerrors "github.com/juju/errors"
 )
 
 // nil bool int8 int32 int64 float32 float64 time.Time
@@ -161,7 +162,7 @@ func encBool(v bool, b []byte) []byte {
 // ::= [x80-xbf]             # -x10 to x3f
 // ::= [xc0-xcf] b0          # -x800 to x7ff
 // ::= [xd0-xd7] b1 b0       # -x40000 to x3ffff
-// com/alibaba/com/caucho/hessian/io/Hessian2Output.java:642 WriteInt
+// hessian-lite/src/main/java/com/alibaba/com/alibaba/com/caucho/hessian/io/Hessian2Output.java:642 WriteInt
 func encInt32(v int32, b []byte) []byte {
 	if int32(INT_DIRECT_MIN) <= v && v <= int32(INT_DIRECT_MAX) {
 		return encByte(b, byte(v+int32(BC_INT_ZERO)))
@@ -184,7 +185,7 @@ func encInt32(v int32, b []byte) []byte {
 // ::= [xf0-xff] b0          # -x800 to x7ff
 // ::= [x38-x3f] b1 b0       # -x40000 to x3ffff
 // ::= x59 b3 b2 b1 b0       # 32-bit integer cast to long
-// com/alibaba/com/caucho/hessian/io/Hessian2Output.java:642 WriteLong
+// hessian-lite/src/main/java/com/alibaba/com/alibaba/com/caucho/hessian/io/Hessian2Output.java:642 WriteLong
 func encInt64(v int64, b []byte) []byte {
 	if int64(LONG_DIRECT_MIN) <= v && v <= int64(LONG_DIRECT_MAX) {
 		return encByte(b, byte(v+int64(BC_LONG_ZERO)))
@@ -426,7 +427,7 @@ func buildMapKey(key reflect.Value, typ reflect.Type) interface{} {
 	}
 
 	// return nil
-	return newCodecError("unsuport key kind " + typ.Kind().String())
+	return jerrors.Errorf("unsuport key kind %s", typ.Kind().String())
 }
 
 func (e *Encoder) encMap(m interface{}) error {