+++ codec/hessian/change_log.md
@@ -6,6 +6,17 @@
 ---
 *本项目其实是从2016/10/21开始进行改造测试的工作了，但2016/10/28才上传到github上，所以dev list只能从29号开始记了*
 
+
+### 2018-04-27 ###
+---
+- 1 improvement: pkg/errors -> juju/errors
+- 2 bug fix:  ReflectResponse 中返回值为struct的时候对返回值（reflect.Value）转换为struct 对象
+
+### 2018-04-26 ###
+---
+- 1 bug fix: encInt64在v值介于 [xd8-xef]   时，应该是 “加上” BC_LONG_ZERO；
+
+
 ### 2017-10-16 ###
 ---
 - 1 添加解析dubbo response相关代码；